+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added loading parameters from a Python file.
 - Added saving parameters to a JSON file.
 - Added loading parameters from a file as a parameter set.
+- Added saving parameter sets to a CSV file.
 - Added saving software versions to a JSON file.
 - Added generating simulation id based on local date and time.
 - Added creating directory structure for simulation.